css/css-animations/animationevent-pseudoelement.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta name="timeout" content="long">
<title>CSS Animations Test: AnimationEvent pseudoElement</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://drafts.csswg.org/css-animations/#interface-animationevent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  div:before {
    width: 100px;
    height: 100px;
    display: block;
    content: " ";
    background-color: blue;
    animation: move 1s;
  }

  @keyframes move {
    to { transform: translate(100px); }
  }
</style>
<div id=div></div>
<script>
  async_test(function(t) {
    div.addEventListener("animationstart", t.step_func(function(evt) {
      assert_true(evt instanceof window.AnimationEvent);
      assert_equals(evt.pseudoElement, "::before");

      t.done();
    }), true);
  }, "pseudoElement of AnimationEvent");
</script>
